+++ lib/components/Button/Button.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import * as classnames from 'classnames/bind';
 import {SpanProps, ButtonProps as AttrButtonProps, Elements as Attr} from '../../Attributes';
+import { StyledElements, ButtonTheme } from '../../Styled';
 const css = classnames.bind(require('./Button.module.scss'));
 
 export interface ButtonComponentType {}
@@ -32,6 +33,8 @@ export interface ButtonProps extends React.Props<ButtonComponentType> {
     className?: string;
 
     attr?: ButtonAttributes;
+    
+    theme?: ButtonTheme;
 }
 
 /**
@@ -45,8 +48,10 @@ export const Button: React.StatelessComponent<ButtonProps> = (props: ButtonProps
         attr={props.attr.icon}
     /> : '';
 
+    const ButtonWrapper = props.theme ? StyledElements.button : Attr.button;
+
     return (
-        <Attr.button
+        <ButtonWrapper
             type={props.type}
             title={props.title}
             className={css('btn', {
@@ -55,10 +60,11 @@ export const Button: React.StatelessComponent<ButtonProps> = (props: ButtonProps
             onClick={props.onClick}
             disabled={props.disabled}
             attr={props.attr.container}
+            theme={props.theme}
         >
             {icon}
             {props.children}
-        </Attr.button>
+        </ButtonWrapper>
     );
 };