+++ lib/components/Masthead/Masthead.tsx
@@ -1,6 +1,9 @@
 import * as React from 'react';
 import * as classnames from 'classnames/bind';
+import styled from 'styled-components';
+
 import { MethodNode } from '../../Common';
+import { StyledElements, ComponentTheme } from '../../Styled';
 import * as InlinePopup from '../InlinePopup';
 import { NavigationProperties } from '../Navigation/Navigation';
 import { ActionTriggerButton, ActionTriggerAttributes, ActionTriggerButtonAttributes } from '../ActionTrigger';
@@ -24,7 +27,7 @@ export interface MastheadSearchItem {
 
     /** The user input value */
     value: string;
-    
+
     /** 
      * For small screen sizes, the search input is collapsed and replaced with a
      * toolbar button. When the button is clicked, the input is shown and occupies
@@ -32,16 +35,16 @@ export interface MastheadSearchItem {
      * width expanded view is shown.
      */
     expanded?: boolean;
-    
+
     /** Event handler to call when the search input should be submitted. */
     onSubmit: React.EventHandler<any>;
-    
+
     /** Event handler to call when the search `value` should be changed. */
     onChange: (newValue: string) => void;
-    
+
     /** Event handler to call when the `expanded` property  should be toggled. */
     onExpand: React.EventHandler<any>;
-    
+
     attr?: TextInputAttributes;
 }
 
@@ -53,6 +56,14 @@ export interface MastheadUserItem {
     displayName: string;
     email: string;
     attr?: InlinePopup.Attributes;
+}
+
+export interface MastheadTheme extends ComponentTheme {
+    colorBackground: string;
+    colorButtonRest: string;
+    colorButtonHover: string;
+    colorTextRest: string;
+    colorTextDisabled: string;
 }
 
 export interface MastheadProperties {
@@ -68,6 +79,7 @@ export interface MastheadProperties {
     };
     toolbarItems?: Array<MastheadToolbarItem>;
     user?: React.ReactNode;
+    theme?: MastheadTheme;
 }
 
 export class Masthead extends React.PureComponent<MastheadProperties> {
@@ -78,17 +90,23 @@ export class Masthead extends React.PureComponent<MastheadProperties> {
         }
         return this.props.toolbarItems.map((item, idx) => {
             const { label, icon, onClick, selected, attr } = item;
-
             return (
                 <li key={idx} className={cx('masthead-toolbar-btn-container', { 'selected-more': this.props.more.selected })}>
-                    < ActionTriggerButton
+                    <ActionTriggerButton
                         label={label}
                         attr={attr}
                         icon={icon}
                         onClick={onClick}
                         className={cx('masthead-toolbar-btn', { 'selected': selected })}
+                        theme={this.props.theme && {
+                            colorRest: this.props.theme.colorBackground,
+                            colorHover: this.props.theme.colorButtonHover,
+                            colorDisabled: this.props.theme.colorTextRest,
+                            colorTextRest: this.props.theme.colorTextRest,
+                            colorTextDisabled: this.props.theme.colorTextDisabled
+                        }}
                     />
-                </li >
+                </li>
             );
         });
     }
@@ -99,12 +117,22 @@ export class Masthead extends React.PureComponent<MastheadProperties> {
             user,
             search,
             more,
-            logo
+            logo,
+            theme,
+            branding
         } = this.props;
+
         const items = this.getToolbarItems();
         const expanded = search && search.expanded;
+
+        // Use style property for banner to speed up initial rendering.
+        const mastheadStyles = theme && {
+            backgroundColor: theme.colorBackground,
+            color: theme.colorTextRest
+        } as React.CSSProperties;
+
         return (
-            <Attr.div key='Masthead' role='banner' className={cx('masthead')}>
+            <Attr.div key='Masthead' role='banner' className={cx('masthead')} style={mastheadStyles}>
                 {navigation &&
                     <InlinePopup.Container
                         expanded={navigation.isExpanded}
@@ -139,6 +167,13 @@ export class Masthead extends React.PureComponent<MastheadProperties> {
                                 icon={'search'}
                                 onClick={search.onExpand}
                                 className={cx('masthead-toolbar-btn')}
+                                theme={theme && {
+                                    colorRest: theme.colorBackground,
+                                    colorHover: theme.colorButtonHover,
+                                    colorDisabled: theme.colorTextRest,
+                                    colorTextRest: theme.colorTextRest,
+                                    colorTextDisabled: theme.colorTextDisabled
+                                }}
                             />
                         </li>}
                         {more && !more.selected && items}
@@ -169,8 +204,8 @@ export class Masthead extends React.PureComponent<MastheadProperties> {
                             {user}
                         </li>}
                     </ul >
-                </Attr.div >
-            </Attr.div >
+                </Attr.div>
+            </Attr.div>
         );
     }
 }