+++ lib/index.ts
@@ -18,3 +18,4 @@ export * from './components/Shell';
 export * from './components/Thumbnail';
 export * from './components/Toggle';
 export * from './Common';
+export * from './Styled';