+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v6.1.0
 - Add logo to Masthead
+- Update Masthead, ActionTriggerButton to support customized styles
 
 ## v6.0.7
 ### Changed